Finished stats page for now

+++ lib/stats_page.dart
@@ -19,9 +19,8 @@ class StatsPage extends StatelessWidget {
             Row(
               children: [
                 Padding(
-                  padding: const EdgeInsets.all(8.0),
-                  child: Text("Time: ${studyData.getTime()}"),
-                ),
+                    padding: const EdgeInsets.all(8.0),
+                    child: Text("Today: ${studyData.getTimeDaily()}")),
                 // space that takes up the rest of the space
                 const Spacer(),
                 Padding(
@@ -31,26 +30,13 @@ class StatsPage extends StatelessWidget {
                     onSelected: (index) {
                       studyData.dropdownChanged(index);
                     },
-                    initialSelection: 0,
+                    initialSelection: studyData.currentSelected,
                   ),
                 ),
               ],
             ),
             SizedBox(height: 20),
-            SfCartesianChart(
-              plotAreaBorderWidth: 0,
-              title: ChartTitle(
-                text: "${studyData.getTimeWeekly()}",
-              ),
-              primaryXAxis: CategoryAxis(
-                majorGridLines: const MajorGridLines(width: 0),
-              ),
-              primaryYAxis: NumericAxis(
-                  axisLine: const AxisLine(width: 0),
-                  labelFormat: '{value}',
-                  majorTickLines: const MajorTickLines(size: 0)),
-              series: studyData.getWeeklySeries(),
-            ),
+            StatsWidget(studyData: studyData),
             Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
@@ -58,7 +44,30 @@ class StatsPage extends StatelessWidget {
                   padding: const EdgeInsets.all(8.0),
                   child: ElevatedButton(
                     onPressed: () {
-                      studyData.deleteStats();
+                      showDialog(
+                          context: context,
+                          builder: (context) {
+                            return AlertDialog(
+                              title: const Text("Reset"),
+                              content: const Text(
+                                  "Are you sure you want to reset your data?"),
+                              actions: [
+                                TextButton(
+                                  onPressed: () {
+                                    Navigator.of(context).pop();
+                                  },
+                                  child: const Text("Cancel"),
+                                ),
+                                TextButton(
+                                  onPressed: () {
+                                    studyData.deleteStats();
+                                    Navigator.of(context).pop();
+                                  },
+                                  child: const Text("Reset"),
+                                ),
+                              ],
+                            );
+                          });
                     },
                     child: const Text("Reset"),
                   ),
@@ -66,24 +75,148 @@ class StatsPage extends StatelessWidget {
                 Padding(
                   padding: const EdgeInsets.all(8.0),
                   child: ElevatedButton(
-                      onPressed: () {
-                        studyData.testMethod();
-                      },
-                      child: const Text("Test")),
-                ),
-                Padding(
-                  padding: const EdgeInsets.all(8.0),
-                  child: ElevatedButton(
-                      onPressed: () {
-                        studyData.testMethod2();
-                      },
-                      child: const Text("Test2")),
+                    onPressed: () {
+                      showDialog(
+                          context: context,
+                          builder: (context) {
+                            return StatefulBuilder(builder:
+                                (BuildContext context, StateSetter setState) {
+                              return AlertDialog(
+                                title: const Text("Manual"),
+                                content: SingleChildScrollView(
+                                  child: ConstrainedBox(
+                                    constraints: const BoxConstraints(
+                                      minHeight: 50,
+                                      maxHeight: 100,
+                                    ),
+                                    child: Column(
+                                      crossAxisAlignment:
+                                          CrossAxisAlignment.start,
+                                      children: [
+                                        ElevatedButton(
+                                          onPressed: () {
+                                            showDatePicker(
+                                              context: context,
+                                              initialDate: DateTime.now(),
+                                              firstDate: DateTime(2019),
+                                              lastDate: DateTime.now(),
+                                            ).then((date) {
+                                              if (date != null) {
+                                                studyData.manualAdd(date);
+                                                setState(() {});
+                                              }
+                                            });
+                                          },
+                                          child: Text(
+                                              "${studyData.manualDate.day}/${studyData.manualDate.month}/${studyData.manualDate.year}"),
+                                        ),
+                                        ElevatedButton(
+                                          onPressed: () {
+                                            showTimePicker(
+                                              context: context,
+                                              initialTime: const TimeOfDay(
+                                                  hour: 0, minute: 0),
+                                              initialEntryMode:
+                                                  TimePickerEntryMode.input,
+                                              helpText: "Time spent studying",
+                                            ).then((time) {
+                                              if (time != null) {
+                                                studyData.manualAddTime(time);
+                                                setState(() {});
+                                              }
+                                            });
+                                          },
+                                          child: Text(
+                                              "${studyData.manualTime.hour.toString().padLeft(2, "0")}:${studyData.manualTime.minute.toString().padLeft(2, "0")}"),
+                                        ),
+                                      ],
+                                    ),
+                                  ),
+                                ),
+                                actions: [
+                                  TextButton(
+                                    onPressed: () {
+                                      Navigator.of(context).pop();
+                                    },
+                                    child: const Text("Cancel"),
+                                  ),
+                                  TextButton(
+                                    onPressed: () {
+                                      studyData.manualAddComplete();
+                                      Navigator.of(context).pop();
+                                    },
+                                    child: const Text("Add"),
+                                  ),
+                                ],
+                              );
+                            });
+                          });
+                    },
+                    child: const Text("Manual"),
+                  ),
                 ),
               ],
             ),
           ],
         ),
       ),
+    );
+  }
+}
+
+class StatsWidget extends StatelessWidget {
+  const StatsWidget({
+    super.key,
+    required this.studyData,
+  });
+
+  final StudyData studyData;
+
+  @override
+  Widget build(BuildContext context) {
+    switch (studyData.currentSelected) {
+      case 0:
+        return Placeholder();
+      case 1:
+        return _buildWeeklyChart();
+      case 2:
+        return _buildMonthlyChart();
+      default:
+        return Placeholder();
+    }
+  }
+
+  Widget _buildWeeklyChart() {
+    return SfCartesianChart(
+      plotAreaBorderWidth: 0,
+      title: ChartTitle(
+        text: "${studyData.getTimeWeekly()}",
+      ),
+      primaryXAxis: CategoryAxis(
+        majorGridLines: const MajorGridLines(width: 0),
+      ),
+      primaryYAxis: NumericAxis(
+          axisLine: const AxisLine(width: 0),
+          labelFormat: '{value}',
+          majorTickLines: const MajorTickLines(size: 0)),
+      series: studyData.getWeeklySeries(),
+    );
+  }
+
+  Widget _buildMonthlyChart() {
+    return SfCartesianChart(
+      plotAreaBorderWidth: 0,
+      title: ChartTitle(
+        text: "${studyData.getTimeMonthly()}",
+      ),
+      primaryXAxis: CategoryAxis(
+        majorGridLines: const MajorGridLines(width: 0),
+      ),
+      primaryYAxis: NumericAxis(
+          axisLine: const AxisLine(width: 0),
+          labelFormat: '{value}',
+          majorTickLines: const MajorTickLines(size: 0)),
+      series: studyData.getMonthlySeries(),
     );
   }
 }
